# Optimize: Filter/map rowan nodes instead of doing multiple lookups