@milianw 2024-10-26
FTR: if we could make progress on https://github.com/rust-analyzer/rowan/pull/119 then this would drastically benefit `slint` too - iterating over the rowan tree is still the main source of allocations and the performance impact is quite measurable (as can be seen from this patch here e.g.)